110.643MHz/1A6E: VT100 mostly fixed!
The VT100 component is mostly fixed now, there is just a minor problem
with where the attributes are place (the first column of pixels
attributes bleeds into the last column). This will most likely be fixed
soon.

The problem was caused by using the next attributes with the previous
character being displayed - it had nothing to do with the VT100 core
itself (which is where most of the complexity lies, it deals with
writing the character/attribute, not reading it).