110.643MHz/1A6E: VGA Controller, Hamming code, Generics
* A VGA controller has been added, this controller needs more work
however as it currently does not work very well (or at all really).
Demonstration code has been added to the project which draws a
red background (incorrectly).
* The Hamming code module has been corrected (extra parity bit has been
calculated correctly now), and SECDEC has been implemented as a
configurable option. The test bench for it has been improved as well,
single and double bit errors can be simulated and the correct behavior
asserted for.
* The continued fight to adopt the new system of generics continues.