103.556MHz/1AEC: Added my own UART module
I have replaced the existing UART module with one of my own design,
which is more flexible (it can be configured at run time, and various
other aspects can be set which the previous one could not - it also has
the same license as the rest of the project).

There is still things to do in relation to the new UART module, some new
configuration registers need adding and the test bench changing. Once
done, this UART will be integrated into 'util.vhd'.

Some words have been simplified in the 'h2.fth' project, but overall the
size of the image has increased as test words have been added (such as a
word that tests the VT100 display routines.