103.556MHz/1B10: Selectable font, UART Config
The default font has been changed to be terminus, the actual characters
used are not optimal and should be chosen from a better set, the font
itself can be selected from with the appropriate ANSI escape sequence.
This expands the size of the block RAM used from 4096 to 8192 (however,
each block RAM is in reality up to 18KiB in size, so there is more space
for extra fonts if needed).

Other improvements include:

* Adding 'delay' to the 'uart.vhd' package.
* Making the UART much more configurable
* Updating the simulators to account for the new registers
* Removing unneeded generics