103.556MHz/1B00: Moved font, ++ VT100, test code
The VT100 terminal emulator has been much improved, and a new attribute
can be applied to the text (whilst not losing any current attributes).
Text can be blinked on and off (using a clock derived from the cursor
blink clock, divided by four), the cursor save and restore command has
been added, along with a few other miscellaneous commands, and a bug in
the relative cursor positioning routines has been fixed (a relative
move of 0 should be treated the same as 1 for those functions, it was
not being treated that way).

* UART baud initial value can be set by a generic, which is nice to
have, but not needed as it can be set from a control register. The other
options cannot be set at the moment and default to 8 bits, 1 stop bit,
no parity.
* Special cases for majority logic gates of 1, 2 and 3 have been added
to the 'majority' module.
* The font directory has moved.
* 'vga.vhd' converted to a UTF-8 file, fixed some invalid characters.
* Formatting changed 'port(' -> 'port ('
* Added extra VT100 terminal codes to simulator.
* Instead of commenting out code, generate functions are used.
  * This applies to the new VGA code.
  * And to adding a FIFO to the UART (which almost works).
* Some extra comments have been added.

The next major problems with the project will be:
* Making the UART more reliable, currently we get nonsense if we are not
processing received characters quick enough and we spend a lot of time
waiting around for transmission. There is a rate that we will not be
able to keep up with when it comes to processing text, however we should
never be receiving invalid data, which we are, if we cannot keep up.
* The UART RX module does not work with certain values for the baud
generator (specifically at 115200, the value that needs to go into the
clock register is '54', however it does not work when given that value,
50-53 and 55 have been tried and work, 54 does not). For TX, '54' works.
* Scrolling should be handled by the VT100 terminal, currently it just
erases everything on the screen. This is not optimal.