103.556MHz/1AFE: Basic VT100 scrolling
Instead of blanking the screen after reaching the end of the screen,
basic scrolling has been added. There are some minor bugs, for example
not all of the characters are always cleared as the screen moves up.

Scrolling back up could be implemented; you could only go as far back as
the base pointer will allow.

There are also more warnings and 'INFO' statements in the generated
reports, which are always annoying, these should be removed before
merging this branch into master.