109.636MHz/1B00: Many minor improvements
* Initial UART settings can be set by a generic.
* UART RX now only starts when there is a transition from high to low,
there are still problems (a frame error does not propagate to the rest
of the system, or is not detected), but this means when the UART line is
driven low then we do not get a series of invalid characters.
* The simulator now simulates the scrolling behavior of the VT100, along
with some more of the SGR codes the hardware supports.
* Some warnings have been fixed.
* The 'turn of reverse video' SGR has been fixed, it was $27 when it
should have been decimal 27, or $1B.
* We now erase three rows of text ahead of moving the base pointer text,
this prevents errant text from previous lines being displayed.
* An aid for test vector generation for the VT100 has been added, which
is a small GForth program that can be used to generate text with lots of
ANSI escape sequences in it.

NOTES:

Many of the more serious bugs in the VT100 have been fixed, however
there are still differences in how the simulator behaves, and how the
terminal emulators on my computer behave, which will need to be fixed.

Some of the bugs include:

* The AT-XY command does not seem to work after issue an erase display
command to the VT100, this likely due to the VT100 being too busy to
respond. This only appears when the UART FIFO is enabled, or when UART
output is disable (output only going to the VT100). This means it is
giving commands to the VT100 much quicker.
* The UART FIFO RX seems to work fine, although there are the same
problems with RX as there are without the FIFO (that is, we receive
invalid characters with high data rates). For the TX however, a
character seems to get stuck in the FIFO, it comes out, but then the
next character gets stuck, and so on. No data seems to be lost, but it
is annoying.
* There are some other VT100 differences between the hardware and how it
should behave, the test suite should help categorize and find the
issues.
* Other worrying things include the fact that the board seems to need
reprogramming multiple times before the system gets into a stable state.
Once in that stable state is appears to be very steady and stable, but
it can require multiple times before it becomes stable. Symptoms include
not booting, '? -4 stack underflow' being printed, only 'eForth V' being
printed, or reading from flash failing. I am not sure if the problem is
with my design, however reprogramming seems to fix the issue when it
does occur, with no other design changes needed.