111.949MHz/1A7A: UART FIFO working
The UART FIFO is now working. This requires some changes to the UART
driver, and also how the PS/2 Keyboard interface works. It also has
introduced some difficulties in the simulator, the simulator now needs
to peek at the input stream in order to simulate the PS/2 and UART input
correctly, which is fine for the GUI application, but not so for the CLI
application. I may change how the FIFO works again so it is easier to
simulate. The problem is whether we read a character first and then
trigger a read enabled, or whether we do a read enable then read a
character.