110.358MHz/1A0A: Smaller image, remove some todos
The eForth image has been made smaller, with the removal of certain
unneeded words and the creation of a few new ones. Some new assembly
routines have also been added. This was aided partially by analyzing the
'embed.fth' file textually for bigrams, trigrams and instances of words
that only occurred once and that are not visible to the user.

Some unneeded todos have been removed as well.