110.358MHz/1A0C: Bug fix in immediate
The word 'immediate' was broken in the target Forth due to an
optimization gone wrong. The word 'toggle' should not have been moved to
where it had been moved.

Others changes include some white-space changes and adding a break into
the meta-compiler virtual machine.

Direction of project:
* Keep improving the meta-compiler and changing it so it resembles the
target machine more and more until the system is capable of self-hosting
* Add CGA graphics capability
* Clean up the 'h2.c' interface and header, especially around how 'main'
is defined/used.
* Improve the GUI simulator with textures and perhaps by using SDL.